Organize Facility Locator search form components
