+++ src/applications/facility-locator/components/search-form/autosuggest/DownshiftCaret.jsx
