+++ src/applications/facility-locator/components/search-form/autosuggest/DownshiftClear.jsx
