+++ src/applications/facility-locator/components/search-form/autosuggest/InputWithClear.jsx
