+++ src/applications/facility-locator/components/search-form/autosuggest/StateReducer.js
